Improve naming of the poms and introduce build pom
remove unused ets-build-parent (ets-poms-build is the pom in the root
directory). Move the original root pom (which is a parent pom) to the
ets-poms-root-parent folder.

This is the preparation for configuring the distribution management and
scmurls of the ets-poms project - they must not be included in the
parent poms of the ets poms because they must not be inherited.