Add scm declaration to each pom.
Maven auto appends paths to scm urls in child poms, see MNG-3244 and
MNG-2290. We have to declare the scm connection in the root parent and
it would be enough to declare it in the aggregator pom for releasing
too, but to keep it consistent I declare it in every pom now correctly.