Merge dependency-pom with module parent pom
When using a seperate dependencies pom with scope import, you are not
able to overwrite any properties defining versions of a library because
the dependencies pom is not inherited by your pom (not merged to an
effetive pom). If we declare the dependency management in the module
parent pom (which is the parent of our services and libs), we are able
to overwrite library versions simply by redefining the corresponding
property (we extracted all versions to properties).

This may be mandatory in cases where we want to compile the library
against another scala version (or other edge cases).

Decision is not to use the dependency management import in future.